ROVER-69 Unify supergraph_config parsing
We do this in both supergraph commands and dev
and having it in two places is not helping.
