+++ src/command/supergraph/compose/do_compose.rs
@@ -1,6 +1,7 @@
 use std::{fs::File, io::Write, process::Command, str};
 
 use anyhow::{anyhow, Context};
+use apollo_federation_types::config::FederationVersion::LatestFedTwo;
 use apollo_federation_types::config::SupergraphConfig;
 use apollo_federation_types::{
     build::BuildResult,
@@ -10,10 +11,11 @@ use camino::Utf8PathBuf;
 use clap::Parser;
 use serde::Serialize;
 
+use rover_client::shared::GraphRef;
 use rover_client::RoverClientError;
 use rover_std::{Emoji, Style};
 
-use crate::command::supergraph::resolve_supergraph_yaml;
+use crate::utils::supergraph_config::get_supergraph_config;
 use crate::utils::{client::StudioClientConfig, parsers::FileDescriptorType};
 use crate::{
     command::{
@@ -24,22 +26,45 @@ use crate::{
     RoverError, RoverErrorSuggestion, RoverOutput, RoverResult,
 };
 
-#[derive(Debug, Clone, Serialize, Parser)]
+#[derive(Debug, Serialize, Parser)]
 pub struct Compose {
+    #[clap(flatten)]
+    opts: SupergraphComposeOpts,
+}
+
+#[derive(Debug, Serialize, Parser)]
+pub struct SupergraphComposeOpts {
+    #[clap(flatten)]
+    pub plugin_opts: PluginOpts,
+
     /// The relative path to the supergraph configuration file. You can pass `-` to use stdin instead of a file.
     #[serde(skip_serializing)]
     #[arg(long = "config")]
     supergraph_yaml: FileDescriptorType,
 
-    #[clap(flatten)]
-    opts: PluginOpts,
+    /// A [`GraphRef`] that is accessible in Apollo Studio.
+    /// This is used to initialize your supergraph with the values contained in this variant.
+    ///
+    /// This is analogous to providing a supergraph.yaml file with references to your graph variant in studio.
+    ///
+    /// If used in conjunction with `--supergraph-config`, the values presented in the supergraph.yaml will take precedence over these values.
+    #[arg(long = "graph-ref")]
+    graph_ref: Option<GraphRef>,
+
+    /// The version of Apollo Federation to use for composition
+    #[arg(long = "federation-version")]
+    federation_version: Option<FederationVersion>,
 }
 
 impl Compose {
     pub fn new(compose_opts: PluginOpts) -> Self {
         Self {
-            supergraph_yaml: FileDescriptorType::File("RAM".into()),
-            opts: compose_opts,
+            opts: SupergraphComposeOpts {
+                plugin_opts: compose_opts,
+                supergraph_yaml: FileDescriptorType::File("RAM".into()),
+                graph_ref: None,
+                federation_version: Some(LatestFedTwo),
+            },
         }
     }
 
@@ -52,6 +77,7 @@ impl Compose {
         let plugin = Plugin::Supergraph(federation_version.clone());
         if federation_version.is_fed_two() {
             self.opts
+                .plugin_opts
                 .elv2_license_accepter
                 .require_elv2_license(&client_config)?;
         }
@@ -60,14 +86,14 @@ impl Compose {
         let install_command = Install {
             force: false,
             plugin: Some(plugin),
-            elv2_license_accepter: self.opts.elv2_license_accepter,
+            elv2_license_accepter: self.opts.plugin_opts.elv2_license_accepter,
         };
 
         // maybe do the install, maybe find a pre-existing installation, maybe fail
         let plugin_exe = install_command.get_versioned_plugin(
             override_install_path,
             client_config,
-            self.opts.skip_update,
+            self.opts.plugin_opts.skip_update,
         )?;
         Ok(plugin_exe)
     }
@@ -80,13 +106,16 @@ impl Compose {
         eprintln!(
             "{}resolving SDL for subgraphs defined in {}",
             Emoji::Hourglass,
-            Style::Path.paint(self.supergraph_yaml.to_string())
+            Style::Path.paint(self.opts.supergraph_yaml.to_string())
         );
-        let mut supergraph_config = resolve_supergraph_yaml(
-            &self.supergraph_yaml,
+        let mut supergraph_config = get_supergraph_config(
+            &self.opts.graph_ref,
+            &Some(self.opts.supergraph_yaml.clone()),
+            &self.opts.federation_version.clone().unwrap_or(LatestFedTwo),
             client_config.clone(),
-            &self.opts.profile,
-        )?;
+            &self.opts.plugin_opts.profile,
+        )?
+        .unwrap();
         self.compose(override_install_path, client_config, &mut supergraph_config)
     }
 
@@ -187,129 +216,11 @@ impl Compose {
 #[cfg(test)]
 mod tests {
     use std::convert::TryFrom;
-    use std::fs;
 
-    use assert_fs::TempDir;
     use rstest::rstest;
     use speculoos::assert_that;
 
-    use houston as houston_config;
-    use houston_config::Config;
-
-    use crate::options::ProfileOpt;
-    use crate::utils::client::ClientBuilder;
-
     use super::*;
-
-    fn get_studio_config() -> StudioClientConfig {
-        let tmp_home = TempDir::new().unwrap();
-        let tmp_path = Utf8PathBuf::try_from(tmp_home.path().to_path_buf()).unwrap();
-        StudioClientConfig::new(
-            None,
-            Config::new(Some(&tmp_path), None).unwrap(),
-            false,
-            ClientBuilder::default(),
-        )
-    }
-
-    #[rstest]
-    fn it_errs_on_invalid_subgraph_path() {
-        let raw_good_yaml = r#"subgraphs:
-  films:
-    routing_url: https://films.example.com
-    schema:
-      file: ./films-do-not-exist.graphql
-  people:
-    routing_url: https://people.example.com
-    schema:
-      file: ./people-do-not-exist.graphql"#;
-        let tmp_home = TempDir::new().unwrap();
-        let mut config_path = Utf8PathBuf::try_from(tmp_home.path().to_path_buf()).unwrap();
-        config_path.push("config.yaml");
-        fs::write(&config_path, raw_good_yaml).unwrap();
-        assert!(resolve_supergraph_yaml(
-            &FileDescriptorType::File(config_path),
-            get_studio_config(),
-            &ProfileOpt {
-                profile_name: "profile".to_string()
-            }
-        )
-        .is_err())
-    }
-
-    #[rstest]
-    fn it_can_get_subgraph_definitions_from_fs() {
-        let raw_good_yaml = r#"subgraphs:
-  films:
-    routing_url: https://films.example.com
-    schema:
-      file: ./films.graphql
-  people:
-    routing_url: https://people.example.com
-    schema:
-      file: ./people.graphql"#;
-        let tmp_home = TempDir::new().unwrap();
-        let mut config_path = Utf8PathBuf::try_from(tmp_home.path().to_path_buf()).unwrap();
-        config_path.push("config.yaml");
-        fs::write(&config_path, raw_good_yaml).unwrap();
-        let tmp_dir = config_path.parent().unwrap().to_path_buf();
-        let films_path = tmp_dir.join("films.graphql");
-        let people_path = tmp_dir.join("people.graphql");
-        fs::write(films_path, "there is something here").unwrap();
-        fs::write(people_path, "there is also something here").unwrap();
-        assert!(resolve_supergraph_yaml(
-            &FileDescriptorType::File(config_path),
-            get_studio_config(),
-            &ProfileOpt {
-                profile_name: "profile".to_string()
-            }
-        )
-        .is_ok())
-    }
-
-    #[rstest]
-    fn it_can_compute_relative_schema_paths() {
-        let raw_good_yaml = r#"subgraphs:
-  films:
-    routing_url: https://films.example.com
-    schema:
-      file: ../../films.graphql
-  people:
-    routing_url: https://people.example.com
-    schema:
-        file: ../../people.graphql"#;
-        let tmp_home = TempDir::new().unwrap();
-        let tmp_dir = Utf8PathBuf::try_from(tmp_home.path().to_path_buf()).unwrap();
-        let mut config_path = tmp_dir.clone();
-        config_path.push("layer");
-        config_path.push("layer");
-        fs::create_dir_all(&config_path).unwrap();
-        config_path.push("config.yaml");
-        fs::write(&config_path, raw_good_yaml).unwrap();
-        let films_path = tmp_dir.join("films.graphql");
-        let people_path = tmp_dir.join("people.graphql");
-        fs::write(films_path, "there is something here").unwrap();
-        fs::write(people_path, "there is also something here").unwrap();
-        let subgraph_definitions = resolve_supergraph_yaml(
-            &FileDescriptorType::File(config_path),
-            get_studio_config(),
-            &ProfileOpt {
-                profile_name: "profile".to_string(),
-            },
-        )
-        .unwrap()
-        .get_subgraph_definitions()
-        .unwrap();
-        let film_subgraph = subgraph_definitions.first().unwrap();
-        let people_subgraph = subgraph_definitions.get(1).unwrap();
-
-        assert_eq!(film_subgraph.name, "films");
-        assert_eq!(film_subgraph.url, "https://films.example.com");
-        assert_eq!(film_subgraph.sdl, "there is something here");
-        assert_eq!(people_subgraph.name, "people");
-        assert_eq!(people_subgraph.url, "https://people.example.com");
-        assert_eq!(people_subgraph.sdl, "there is also something here");
-    }
 
     #[rstest]
     #[case::simple_binary("a/b/c/d/supergraph-v2.8.5", "v2.8.5")]
