+++ src/utils/mod.rs
@@ -3,6 +3,8 @@ pub mod env;
 pub mod parsers;
 pub mod pkg;
 pub mod stringify;
+#[cfg(feature = "composition-js")]
+pub mod supergraph_config;
 pub mod table;
 pub mod telemetry;
 pub mod version;
